[PFS 227] nil out entire branch (#9944)
This PR is to nil out branch field in responses. This is done in
interceptors.
Motivated by need to reduce confusion about relationship between Commits
and Branches.

Most of the changes in this PR are to fix tests. 90%+ of failed tests
are due to nil pointer problem (branch field is nil). A common fix is:
we resolve commit by its id instead of by its branch.