Add infrastructure for testing against real logs; add some real logs and a test (#10010)
This adds some real logs to the logs test. We can accept any
loki-logs.txt from a debug dump in testdata, and it will be available to
the tests.

I got to the bottom of Loki's retention, and made it accept 100 years
worth of logs, so we can freely ingest old logs now. There was some code
to tweak Loki's retention already, but it didn't work (you could only
send logs that were 2 hours and 40 minutes older than the current time).
Now it works, so I removed those options.

I started writing a table-driven test to fetch logs and ensure that we
get the expected values back. In the next PR I'll increase the test
coverage over real logs and move some existing tests to this framework.
The want = verbatim log lines is a very tedious form of tests; we'll
have some of those, but for testing things like paging, going outside
Loki's time range, etc., we'll probably add some assertions to check the
"shape" of the returned data instead of the exact content.