Add storage service to pachctl misc grpc (#10048)
I discovered that the storage service wasn't getting resolved when
running `pachctl misc grpc`. It turns out that the services are
statically configured. I wonder if we can autogenerate this file by
searching for `protoreflect.FileDescriptors` in our code and then adding
lines here for each one of them. That could be a fun follow up.